+++ lib/roadie/filesystem_provider.rb
@@ -34,6 +34,9 @@ module Roadie
       end
     end
 
+    def to_s() inspect end
+    def inspect() "#<#{self.class} #@path>" end
+
     private
     def build_file_path(name)
       raise InsecurePathError, name if name.include?("..")