+++ Changelog.md
@@ -8,6 +8,7 @@
     * Supports `1.5.0`...`1.7.0` now instead of `1.6.0`...`1.7.0`. Some people out there are stuck on this older version of Nokogiri, and I don't want to leave them out.
   * Output better errors when no assets can be found.
     * The error will now show which providers were tried and in which order, along with the error message from the specific providers.
+    * `Roadie::FilesystemProvider` shows the given path when inspected.
   * `data-roadie-ignore` attributes will now be removed from markup; hiding "development markers" in the final email.
 * **Deprecations:**
   * `Roadie::Stylesheet#each_inlinable_block` is now deprecated. You can iterate and filter the `blocks` at your own discresion.